requirements.txt
# PyTorch stack (CUDA 12.1)
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121
--extra-index-url https://download.pytorch.org/whl/cu121

# Ultralytics (YOLO11)
ultralytics==8.3.0

# Ciencia de datos / utilidades
numpy==1.26.4
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
scikit-learn==1.5.2
tqdm==4.66.5

# Visión por computadora
opencv-python==4.10.0.84

# Notebooks
jupyterlab==4.2.5
ipykernel==6.29.5
notebook==7.2.2

# Otros requerimientos de tu script
pyyaml==6.0.2
requests==2.32.3
psutil==5.9.8
